Rename includes for consistency
