+++ ferry/includes/transform_import.py
@@ -2,9 +2,7 @@
 Functions for importing information into tables.
 Used by /ferry/transform.py.
 """
-import csv
 from collections import Counter
-from io import StringIO
 from itertools import combinations
 from typing import List
 
@@ -22,15 +20,6 @@ from ferry.includes.utils import (
 
 # maximum question divergence to allow
 QUESTION_DIVERGENCE_CUTOFF = 32
-
-
-class DatabaseImportError(Exception):
-    """
-    Object for import exceptions.
-    """
-
-    # pylint: disable=unnecessary-pass
-    pass
 
 
 def resolve_cross_listings(merged_course_info):
@@ -807,52 +796,3 @@ def import_evaluations(
         evaluation_statistics,
         evaluation_questions,
     )
-
-
-def copy_from_stringio(conn, table, table_name: str):
-    """
-    Save DataFrame in-memory and migrate
-    to database with copy_from().
-
-    Parameters
-    ----------
-    conn: psycopg2 connection object
-
-    table: DataFrame to import
-
-    table_name: name of target table
-
-    Returns
-    -------
-    """
-
-    # create in-memory buffer for DataFrame
-    buffer = StringIO()
-
-    csv_kwargs = dict(
-        index_label="id",
-        header=False,
-        index=False,
-        sep="\t",
-        quoting=csv.QUOTE_NONE,
-        escapechar="\\",
-        na_rep="NULL",
-    )
-
-    table.to_csv(buffer, **csv_kwargs)
-
-    buffer.seek(0)
-
-    cursor = conn.cursor()
-
-    try:
-        cursor.copy_from(
-            buffer, table_name, columns=table.columns, sep="\t", null="NULL"
-        )
-    except Exception as error:
-        conn.rollback()
-        cursor.close()
-        raise DatabaseImportError from error
-
-    print(f"Successfully copied {table_name}")
-    cursor.close()
